Convert to lat/lng using transform function
If you use the transform from the Rasterio image, that will only convert from
pixel to imagery coordinates, but not back to lat/lng which is needed for GeoJSON
output. By allowing a custom transform function we can support this and also
relax the assumption that the imagery comes from Raster.